Fix targetDir in create-vitebook
On windows, with Git Bash, I was receiving the overwrite prompt when running `yarn create vitebook` in my project root. Since the two checks that matter are `targetDir !== process.cwd()` and `fs.existsSync(targetDir)`, I believe the cause of the problem is a Windows and/or Git Bash path quirk.